Revert "Change value type to float to have full number"
This reverts commit 31b7b9d05128a295f62124237606bf67651ec1b8.

+++ wp-content/themes/toomoney/inc/update-rates.php
@@ -72,7 +72,7 @@ function convertItRapid($t,$f){
         $result = $err;
     } else {
         $json = json_decode($response,true);
-        $result = (float)$json->rates[0]->$t->rate;
+        $result = (int)$json->rates[0]->$t->rate;
     }
     return $result;
 }
